+++ kitty/options/parse.py
@@ -3,8 +3,8 @@
 # isort: skip_file
 import typing
 from kitty.conf.utils import (
-    merge_dicts, positive_float, positive_int, to_bool, to_cmdline, to_color, to_color_or_none,
-    unit_float
+    merge_dicts, positive_float, positive_int, python_string, to_bool, to_cmdline, to_color,
+    to_color_or_none, unit_float
 )
 from kitty.options.utils import (
     action_alias, active_tab_title_template, allow_hyperlinks, bell_on_tab, box_drawing_scale,
@@ -1302,7 +1302,7 @@ class Parser:
         ans['url_color'] = to_color(val)
 
     def url_excluded_characters(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
-        ans['url_excluded_characters'] = str(val)
+        ans['url_excluded_characters'] = python_string(val)
 
     def url_prefixes(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['url_prefixes'] = url_prefixes(val)